+++ lib/cldr/plug/plug_accept_language.ex
@@ -1,27 +1,11 @@
 if Code.ensure_loaded?(Plug) do
   defmodule Cldr.Plug.AcceptLanguage do
     @moduledoc """
-    A Phoenix plug that is used to parse the `HTTP Accept-Language header` and
-    set the `Cldr` locale appropriately.
+    Parses the accept-language header if one is available and sets
+    `conn.private[:cldr_locale]` accordingly.  The locale can
+    be later retrieved by `Cldr.Plug.AcceptLanguage.get_cldr_locale/1`
 
-    The language accept header [defines](http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html)
-    a locale as having up to 8 characters, with an optional dash ('-')
-    and up to 8 additional characters for the region code.  This plug extends that
-    syntax to include a more complete form of a locale string as defined by the
-    [Unicode Consortium](http://unicode.org).
-
-    ## Cldr.Plug.Locale support
-
-    `Cldr.Plug.Locale` will support parsing a locale identifier using the
-    following formats:
-
-    Locale               | Meaning
-    -------------------- | -----------------------
-    en                   | Language "en"
-    en-GB                | Language "en" with region subtag "GB".
-    en-Latn-AU           | Language "en" with script "latn" and regional subtag "AU"
-    en-AU-u-cu-usd       | Language "en", region subtag "AU", currency unit "USD"
-    en-CA-u-cu-use-hc-12 | Multiple extensions can be specified
+    There are no configuration options for this plug.
     """
 
     import Plug.Conn
@@ -29,13 +13,22 @@ if Code.ensure_loaded?(Plug) do
 
     @language_header "accept-language"
 
-    def init(default) do
-      default
+    def init(any) do
+      Logger.warn "#{__MODULE__} does not support configuration options. " <>
+                  "Please remove #{inspect any} from the plug invokation."
+      init()
+    end
+
+    def init do
+      nil
     end
 
     def call(conn, _default) do
-      accept_language = get_req_header(conn, @language_header)
-      put_private(conn, :cldr_locale, best_match(accept_language))
+      case get_req_header(conn, @language_header) do
+        [accept_language] -> put_private(conn, :cldr_locale, best_match(accept_language))
+        [accept_language | _] -> put_private(conn, :cldr_locale, best_match(accept_language))
+        [] -> put_private(conn, :cldr_locale, nil)
+      end
     end
 
     def best_match(nil) do
@@ -52,6 +45,9 @@ if Code.ensure_loaded?(Plug) do
       end
     end
 
+    @doc """
+    Return the locale set by `Cldr.Plug.AcceptLanguage`
+    """
     def get_cldr_locale(conn) do
       conn.private[:cldr_locale]
     end