+++ mix/support/normalize/normalize_datetime.ex
@@ -1,9 +1,6 @@
 # credo:disable-for-this-file
 defmodule Cldr.Normalize.DateTime do
-  @moduledoc """
-  Takes the date part of the locale map and transforms the formats into a more easily
-  processable structure that is then stored in map managed by `Cldr.Locale`
-  """
+  @moduledoc false
 
   def normalize(content, locale) do
     content