+++ mix/support/normalize/normalize_list.ex
@@ -1,8 +1,6 @@
 defmodule Cldr.Normalize.List do
-  @moduledoc """
-  Takes the list part of the locale map and transforms the formats into a more easily
-  processable structure that is then stored in map managed by `Cldr.Locale`
-  """
+  @moduledoc false
+
   alias Cldr.Substitution
 
   def normalize(content, locale) do